requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
scikit-learn 
onnxruntime
